requirements.txt
flask-sqlalchemy
flask-admin
email_validator
mysqlclient
WTForms-Alchemy
